Import evermono-ttf-6.2.1 as fonts/evermono-ttf.
Everson Mono is a simple, elegant, monowidth font. I started
designing it in 1994 primarily to make glyphs available to support
the non-Han characters in Unicode and ISO/IEC 10646-1, though I
hope that users may find it a pleasant alternative to Courier and
Monaco for general purposes, e-mail, and so forth. I have found it
quite legible at sizes as small as 4 points. It is lighter and a
bit looser than Courier.