Update to Asterisk 1.8.21.0: this is a bug fix update
The Asterisk Development Team has announced the release of Asterisk 1.8.21.0.

The release of Asterisk 1.8.21.0 resolves several issues reported by the
community and would have not been possible without your participation.
Thank you!

The following is a sample of the issues resolved in this release:

* --- Fix issue where chan_mobile fails to bind to first available port

* --- Fix station ringback; trunk hangup issues in SLA

* --- Fix Queue Log Reporting Every Call COMPLETECALLER With "h"
      Extension Present

* --- Fix Record-Route parsing for large headers.

* --- Fix AMI redirect action with two channels failing to redirect
      both channels.

For a full list of changes in this release, please see the ChangeLog:

http://downloads.asterisk.org/pub/telephony/asterisk/ChangeLog-1.8.21.0

Thank you for your continued support of Asterisk!