Import p5-Log-Dispatch-Perl-0.04 as devel/p5-Log-Dispatch-Perl.
The "Log::Dispatch::Perl" module offers a logging alternative using
standard Perl core functions.  It allows you to fall back to the
common Perl alternatives for logging, such as "warn" and "cluck".
It also adds the possibility for a logging action to halt the
current environment, such as with "die" and "croak".