Update to 9861: Remove spurious "unfinished/" that appeared at the start of the unfinished puzzles' gamedesc.txt lines due to a Perl error. Ahem.
Add comment to remind myself (and anyone else) how to refresh the XFCE
menu.

Script to read the new gamedesc.txt and create .desktop files. My
immediate intention is to run this locally so that my XFCE main menu
acquires shortcuts for my locally compiled puzzle binaries, but I
expect the script could probably be adapted for systemwide
distribution use if any distribution hasn't already done this job in
their own way by now.

Rename wingames.lst to gamedesc.txt, and add a couple of extra fields
to it giving each game's "internal" name (as seen in the source file,
.R etc) and also a brief description of the game. The idea of the
latter is that it should be usable as a comment field in .desktop
files and similar.