Update lua-lpeg to version 0.12.
What is LPeg? LPeg is a new (well, not that new now) pattern-matching
library for Lua, based on Parsing Expression Grammars (PEGs).

What is new in version 0.12?
  + no "unsigned short" limit for pattern sizes
  + mathtime captures considered nullable  (*)
  + some bugs and compiler warnings fixed

Where to get it?

  http://www.inf.puc-rio.br/~roberto/lpeg/lpeg-0.12-rc1.tar.gz


(*) Although this change can be considered logically a bug fix, it can
lead to some old patterns being refused with a "loop body may accept
empty string" error. It should be easy to fix these patterns.